refactor: update component interface story, styles
