+++ src/shared/ui/table/index.tsx
@@ -1,1 +1,5 @@
-export * from './table'
+export * from './table-body'
+export * from './table-cell'
+export * from './table-head'
+export * from './table-root'
+export * from './table-row'
