+++ src/shared/ui/table/table-body.tsx
@@ -2,12 +2,6 @@ import { ComponentPropsWithoutRef, ElementRef, forwardRef } from 'react'
 
 type Props = ComponentPropsWithoutRef<'tbody'>
 
-export const TableBody = forwardRef<ElementRef<'tbody'>, Props>(
-  ({ children, className, ...rest }, ref) => {
-    return (
-      <tbody ref={ref} {...rest} className={className}>
-        {children}
-      </tbody>
-    )
-  }
-)
+export const TableBody = forwardRef<ElementRef<'tbody'>, Props>(({ ...props }, ref) => {
+  return <tbody ref={ref} {...props} />
+})
